Moved LLModel::prompt to llmodel_shared.cpp
